requirements.txt
aiogram
SpeechRecognition
python-dotenv